Unify naming of classes and events
