+++ core-bundle/src/Resources/config/listener.yml
@@ -397,12 +397,12 @@ services:
         tags:
             - { name: kernel.event_listener, event: kernel.request, method: onKernelRequest }
 
-    contao.listener.generate_descriptor:
-        class: Contao\CoreBundle\EventListener\GenerateDescriptorListener
+    contao.listener.descriptor_generation:
+        class: Contao\CoreBundle\EventListener\Descriptor\DescriptorGenerationListener
         tags:
-            - { name: kernel.event_listener, event: contao.generate_descriptor, method: onDescriptorGeneration }
+            - { name: kernel.event_listener, event: contao.descriptor_generation, method: onDescriptorGeneration }
 
-    contao.listener.generate_content_descriptor:
-        class: Contao\CoreBundle\EventListener\GenerateContentDescriptorListener
+    contao.listener.content_descriptor_generation:
+        class: Contao\CoreBundle\EventListener\Descriptor\ContentDescriptorGenerationListener
         tags:
-            - { name: kernel.event_listener, event: contao.generate_descriptor, method: onDescriptorGeneration }
+            - { name: kernel.event_listener, event: contao.descriptor_generation, method: onDescriptorGeneration }
